fix(web): enforce stricter type definitions
