+++ web/src/routes/+layout.svelte
@@ -3,6 +3,7 @@
 	import 'material-icons/iconfont/material-icons.css'
 	import { serverStatus, endpoint, files, tasks } from '$lib/stores'
 	import { fetchAllRemoteItems } from '$lib/utils/items'
+	import type { FileItem } from '$lib/types'
 	import { get } from 'svelte/store'
 
 	const { children } = $props()
@@ -13,16 +14,12 @@
 		if (!wasOnline && $serverStatus.online) {
 			const ep = get(endpoint)
 
-			fetchAllRemoteItems({
+			fetchAllRemoteItems<FileItem, FileItem>({
 				endpoint: ep,
 				resource: 'uploads',
 				store: files,
 				transform: (raw) => ({
-					id: raw.id,
-					name: raw.file_name,
-					size: raw.file_size,
-					type: raw.mime_type,
-					time: raw.uploaded_at,
+					...raw,
 					icon: 'cloud_sync',
 					progress: 100,
 				}),
@@ -32,10 +29,11 @@
 				endpoint: ep,
 				resource: 'tasks',
 				store: tasks,
-				transform: (raw) => raw,
+				transform: (r) => r,
 			})
 		}
-		wasOnline = get(serverStatus).online
+
+		wasOnline = $serverStatus.online
 	})
 </script>
